refactor: move db version to /stdlib
