+++ yarn-project/foundation/package.json
@@ -52,8 +52,7 @@
     "./validation": "./dest/validation/index.js",
     "./promise": "./dest/promise/index.js",
     "./string": "./dest/string/index.js",
-    "./message": "./dest/message/index.js",
-    "./database-version": "./dest/database-version/index.js"
+    "./message": "./dest/message/index.js"
   },
   "scripts": {
     "build": "yarn clean && tsc -b",
